Fix more TypeScript errors
